consolidated API decoding routes
